# openapi/aurum.yaml
openapi: 3.0.3
info:
  title: Aurum API
  version: 0.1.0
  description: API for retrieving market intelligence data, managing scenarios, and valuing PPAs.
servers:
  - url: https://api.aurum.local
    description: Local development gateway
  - url: https://api.aurum.prod
    description: Production
security:
  - oauth2: [aurum]
paths:
...
  /v1/scenarios:
    get:
      summary: List scenarios for the authenticated tenant.
      operationId: listScenarios
      security:
        - oauth2: [aurum]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Optional scenario status filter
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque cursor for stable pagination
        - name: since_cursor
          in: query
          required: false
          schema:
            type: string
          description: Resume iteration from a previously returned cursor
      responses:
        '200':
          description: Scenarios returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a new scenario definition with assumptions.
      operationId: createScenario
      security:
        - oauth2: [aurum]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScenarioRequest'
      responses:
        '201':
          description: Scenario created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/scenarios/{id}:
    get:
      summary: Retrieve a scenario definition.
      operationId: getScenario
      security:
        - oauth2: [aurum]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Scenario metadata retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a scenario definition and associated runs.
      operationId: deleteScenario
      security:
        - oauth2: [aurum]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Scenario deleted.
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/scenarios/{id}/run:
    post:
      summary: Trigger a scenario run asynchronously via Kafka.
      operationId: runScenario
      security:
        - oauth2: [aurum]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioRunOptions'
      responses:
        '202':
          description: Scenario run accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioRunResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/scenarios/{id}/runs:
    get:
      summary: List runs for a scenario.
      operationId: listScenarioRuns
      security:
        - oauth2: [aurum]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: Optional run state filter
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque cursor for run pagination
        - name: since_cursor
          in: query
          required: false
          schema:
            type: string
          description: Resume iteration from a previously returned run cursor
      responses:
        '200':
          description: Scenario runs returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioRunListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/scenarios/{id}/runs/{run_id}:
    get:
      summary: Get the status of a scenario run.
      operationId: getScenarioRun
      security:
        - oauth2: [aurum]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Scenario run status returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioRunResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/scenarios/runs/{run_id}/cancel:
    post:
      summary: Cancel a scenario run.
      operationId: cancelScenarioRun
      security:
        - oauth2: [aurum]
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Scenario run cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioRunResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/scenarios/runs/{run_id}/state:
    post:
      summary: Update the state of a scenario run.
      operationId: updateScenarioRunState
      security:
        - oauth2: [aurum]
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
            pattern: ^(QUEUED|RUNNING|SUCCEEDED|FAILED|CANCELLED)$
      responses:
        '200':
          description: Scenario run state updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioRunResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/scenarios/{id}/outputs:
    get:
      summary: List scenario outputs
      operationId: listScenarioOutputs
      security:
        - oauth2: [aurum]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: metric
          in: query
          schema:
            type: string
          required: false
        - name: tenor_type
          in: query
          schema:
            type: string
          required: false
        - name: curve_key
          in: query
          schema:
            type: string
          required: false
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          required: false
          description: Maximum number of rows to return (capped at 500 by the service).
        - name: cursor
          in: query
          schema:
            type: string
          required: false
          description: Opaque cursor for stable pagination.
        - name: since_cursor
          in: query
          schema:
            type: string
          required: false
          description: Alias for `cursor` to resume iteration from a previously returned `next_cursor`.
        - name: prev_cursor
          in: query
          schema:
            type: string
          required: false
          description: Cursor pointing to the previous page; use the `meta.prev_cursor` value from a prior response.
        - name: format
          in: query
          required: false
          description: Response format (set to `csv` to stream results).
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Scenario outputs.
          headers:
            ETag:
              description: Strong entity tag for cache validation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioOutputResponse'
            text/csv:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.aurum.local/oauth2/auth
          tokenUrl: https://auth.aurum.local/oauth2/token
          scopes:
            aurum: Full access to Aurum APIs
...
  responses:
    Unauthorized:
      description: Authentication failed or missing token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Meta:
      type: object
      properties:
        request_id:
          type: string
        query_time_ms:
          type: integer
        next_cursor:
          type: string
          nullable: true
        prev_cursor:
          type: string
          nullable: true
...
    CreateScenarioRequest:
      type: object
      required: [tenant_id, name, assumptions]
      properties:
        tenant_id:
          type: string
        name:
          type: string
        description:
          type: string
        assumptions:
          type: array
          items:
            $ref: '#/components/schemas/ScenarioAssumption'
    ScenarioAssumption:
      type: object
      required: [driver_type, payload]
      properties:
        driver_type:
          type: string
        payload:
          type: object
        version:
          type: string
    ScenarioResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: object
          properties:
            scenario_id:
              type: string
            tenant_id:
              type: string
            name:
              type: string
            description:
              type: string
            status:
              type: string
            assumptions:
              type: array
              items:
                $ref: '#/components/schemas/ScenarioAssumption'
            created_at:
              type: string
              format: date-time
    ScenarioRunOptions:
      type: object
      properties:
        code_version:
          type: string
        seed:
          type: integer
    ScenarioRunResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: object
          properties:
            run_id:
              type: string
            scenario_id:
              type: string
            state:
              type: string
            code_version:
              type: string
            seed:
              type: integer
            created_at:
              type: string
              format: date-time
    ScenarioListResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            $ref: '#/components/schemas/ScenarioResponse/properties/data'
    ScenarioRunListResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            $ref: '#/components/schemas/ScenarioRunResponse/properties/data'
    ScenarioOutputPoint:
      type: object
      required: [scenario_id, curve_key, metric]
      properties:
        scenario_id:
          type: string
        run_id:
          type: string
          nullable: true
        asof_date:
          type: string
          format: date
          nullable: true
        curve_key:
          type: string
        tenor_type:
          type: string
          nullable: true
        contract_month:
          type: string
          format: date
          nullable: true
        tenor_label:
          type: string
          nullable: true
        metric:
          type: string
        value:
          type: number
          nullable: true
        band_lower:
          type: number
          nullable: true
        band_upper:
          type: number
          nullable: true
        attribution:
          type: object
          additionalProperties:
            type: number
          nullable: true
        version_hash:
          type: string
          nullable: true
    ScenarioOutputResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            $ref: '#/components/schemas/ScenarioOutputPoint'
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'